requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
Appium-Python-Client==5.1.1
attrs==25.3.0
certifi==2025.6.15
h11==0.16.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
selenium==4.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
